+++ Pulp/Common/String/StringUtil.h
@@ -59,10 +59,10 @@ namespace strUtil
 	inline const wchar_t* Convert(const char* input, wchar_t(&output)[N]);
 
 	/// Converts a wide-character string into a single-byte character string, and returns the converted string.
-	const char* Convert(const wchar_t* input, char* output, uint32_t outputBytes);
+	const char* Convert(const wchar_t* input, char* output, size_t outputBytes);
 
 	/// Converts a wide-character string into a single-byte character string, and returns the converted string.
-	const wchar_t* Convert(const char* input, wchar_t* output, uint32_t outputBytes);
+	const wchar_t* Convert(const char* input, wchar_t* output, size_t outputBytes);
 
 
 	/// Returns the number of occurrences of a character in a string in the given range.
